Rename technique namespaces
