+++ src/SudokuSolver/Techniques/Factories/FishTechniqueFactory.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using SudokuSolver.Techniques.FishTechniques;
 using SudokuSolver.Techniques.Helpers;
 
 namespace SudokuSolver.Techniques.Factories
