+++ src/SudokuSolver/Techniques/ISolverTechnique.cs
@@ -26,7 +26,15 @@ namespace SudokuSolver.Techniques
         IChangeDescription GetPossibleBoardStateChange(BoardState board);
     }
 
-    public record CombinedTechnique(IReadOnlyCollection<ISolverTechniqueDescription> Techniques)
+    internal class NotFound : ISolverTechniqueDescription
+    {
+        private NotFound() { }
+        public static NotFound Instance { get; } = new NotFound();
+        public string Description => "No change found";
+        public DifficultyLevel DifficultyLevel => DifficultyLevel.Undefined;
+    }
+
+    internal record CombinedTechnique(IReadOnlyCollection<ISolverTechniqueDescription> Techniques)
         : ISolverTechniqueDescription
     {
         public string TechniqueName => "Combined Technique";
