--- src/SudokuSolver/Techniques/NotFound.cs
@@ -1,10 +1,0 @@
-namespace SudokuSolver.Techniques
-{
-    internal class NotFound : ISolverTechniqueDescription
-    {
-        private NotFound() { }
-        public static NotFound Instance { get; } = new NotFound();
-        public string Description => "No change found";
-        public DifficultyLevel DifficultyLevel => DifficultyLevel.Undefined;
-    }
-}
