+++ src/SudokuSolver/Techniques/Subsets/SubsetTechniqueBase.cs
@@ -2,7 +2,7 @@ using System.Collections.Immutable;
 using System.Linq;
 using SudokuSolver.Techniques.Helpers;
 
-namespace SudokuSolver.Techniques.SubsetTechniques
+namespace SudokuSolver.Techniques.Subsets
 {
     internal abstract class SubsetTechniqueBase : CollectionCandidateRemover
     {
